Split rootReducer into playerReducers, engineReducers
Also made a helper file for Grid Objects (just overflow checking functions for now).
